+++ app/helpers/recording_studio_support/application_helper.rb
@@ -17,14 +17,6 @@ module RecordingStudioSupport
       Body.sanitize(body).html_safe
     end
 
-    def support_pages_title
-      RecordingStudioSupport.configuration.pages_title
-    end
-
-    def support_pages_subtitle
-      RecordingStudioSupport.configuration.pages_subtitle
-    end
-
     def support_publish_path(recording)
       engine = publishable_engine_routes
       return if engine.blank? || recording.blank?
@@ -38,6 +30,30 @@ module RecordingStudioSupport
       return if time.blank?
 
       "Updated #{time.to_date.to_fs(:long)}"
+    end
+
+    def support_help_title
+      RecordingStudioSupport.configuration.help_title
+    end
+
+    def support_help_subtitle
+      RecordingStudioSupport.configuration.help_subtitle
+    end
+
+    def support_public_help_title
+      RecordingStudioSupport.configuration.public_help_title
+    end
+
+    def support_public_help_subtitle
+      RecordingStudioSupport.configuration.public_help_subtitle
+    end
+
+    def support_public_help_path
+      if respond_to?(:main_app) && main_app.respond_to?(:public_help_path)
+        return main_app.public_help_path
+      end
+
+      RecordingStudioSupport.configuration.public_pages_path
     end
 
     private