+++ lib/generators/recording_studio_support/install/templates/recording_studio_support_initializer.rb
@@ -3,8 +3,10 @@
 RecordingStudioSupport.configure do |config|
   config.pages_path = "/support"
   config.public_pages_path = "/help"
-  config.pages_title = "Help"
-  config.pages_subtitle = "Answers you can share."
-  config.admin_section_title = "Help"
-  config.admin_section_subtitle = "Pages people use when they get stuck."
+  config.help_title = "Help"
+  config.help_subtitle = "Answers you can share."
+  config.public_help_title = "Help"
+  config.public_help_subtitle = "Answers you can read."
+  config.admin_help_title = "Help"
+  config.admin_help_subtitle = "Pages people use when they get stuck."
 end